change type to signed int not unsigned

+++ apps/engine.c
@@ -49,19 +49,19 @@ const OPTIONS engine_options[] = {
 static int append_buf(char **buf, int *size, const char *s)
 {
     const int expand = 256;
-    unsigned int len = strlen(s) + 1;
+    int len = strlen(s) + 1;
     char *p = *buf;
 
     if (p == NULL) {
         *size = (len + expand - 1) / expand * expand;
         p = *buf = app_malloc(*size, "engine buffer");
     } else {
-        const unsigned int blen = strlen(p);
+        const int blen = strlen(p);
 
         if (blen > 0)
             len += 2 + blen;
 
-        if (len > (unsigned int)*size) {
+        if (len > *size) {
             *size = (len + expand - 1) / expand * expand;
             p = OPENSSL_realloc(p, *size);
             if (p == NULL) {
